+++ src/connectedk8s/HISTORY.rst
@@ -5,7 +5,7 @@ Release History
 1.3.7
 ++++++
 
-* Added DNS and outbound connectivity prechecks in connect command
+* Added heuristics change - Remove detection of Distro and infra
 
 1.3.6
 ++++++